fix all register controller bugs

+++ src/utils/cloudinary.js
@@ -17,10 +17,14 @@ const uploadOnCloudinary= async(localFilePath)=>{
             resource_type:"auto"
         })
         // file has been uploaded successfully
-        console.log("file is uploaded on cloudinary",response.url);
+        // console.log("file is uploaded on cloudinary",response.url);
+        fs.unlinkSync(localFilePath);
         return response;
     }catch(error){
-        fs.unlinkSync(localFilePath)
+        console.log("Cloudinary error",error)
+        if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+            }
         return null;
     }
 }
